Fixed tooltips for certain resolutions, stats half-implemented...
And a few other fixes

+++ Assets/Code/Components/Effects/DamageEffect.cs
@@ -4,7 +4,7 @@
 public class DamageEffect : Effect {
 
 	[Tooltip("Damage inflicted to the target entity every time this effect ticks")]
-	[Range(0, 100)] public float m_damage;
+	[Range(0, 100)] public int m_damage;
 
 	public override void Tick(Entity p_target){ 
 		p_target.Damage(null, m_damage, true, true);
